+++ oid4vci/integration/docker-compose.yml
@@ -34,7 +34,6 @@ services:
         --log-level debug
         --debug-webhooks
         --plugin oid4vci
-        --plugin jwt_vs_json
     healthcheck:
       test: curl -s -o /dev/null -w '%{http_code}' "http://localhost:3001/status/live" | grep "200" > /dev/null
       start_period: 30s
